changed completedDist
