+++ src/components/Accordion.jsx
@@ -14,22 +14,27 @@ function Accordion({ items, onDistribute }) {
     console.log("Accordion items to be displayed are:", items);
   }, [items]);
 
+
+
   return (
     <div className="container mx-auto p-4 grid grid-cols-1 lg:grid-cols-2 gap-4 col-span-2 lg:col-span-1">
       {items.length === 0 ? (
         <p className="text-gray-500 dark:text-grey-300 ">No active listings.</p>
       ) : (
         items.map((item, index) => (
-          <AccordionItem
-            key={index}
-            item={item}
-            index={index}
-            expanded={expandedItem === index}
-            onToggle={handleToggle}
-            onDistribute={onDistribute}
-            ariaControls={`accordion-item-${index}`}
-            ariaExpanded={expandedItem === index}
-          />
+          <div>
+          
+            <AccordionItem
+              key={index}
+              item={item}
+              index={index}
+              expanded={expandedItem === index}
+              onToggle={handleToggle}
+              onDistribute={onDistribute}
+              ariaControls={`accordion-item-${index}`}
+              ariaExpanded={expandedItem === index}
+            />
+          </div>
         ))
       )}
     </div>
